Feature(Window) Timing system remade

+++ src/com/example/robotricochet/entities/game/Wall.java
@@ -33,7 +33,7 @@ public class Wall extends Entity implements BoardObject {
     }
 
     @Override
-    public void update(double delta) {
+    public void update(float delta) {
 
     }
